@@ -1,11 +1,11 @@
-import {VgVolume} from "./vg-volume";
-import {VgAPI} from "../../core/services/vg-api";
-import {ElementRef} from "@angular/core";
+import {VgVolume} from './vg-volume';
+import {VgAPI} from '../../core/services/vg-api';
+import {ElementRef} from '@angular/core';
 
 describe('Volume control', () => {
-    let vgVol:VgVolume;
-    let ref:ElementRef;
-    let api:VgAPI;
+    let vgVol: VgVolume;
+    let ref: ElementRef;
+    let api: VgAPI;
 
     beforeEach(() => {
         ref = {
@@ -23,8 +23,8 @@ describe('Volume control', () => {
     it('Should have isDragging set to false initially', () => {
         expect(vgVol.isDragging).toBe(false);
     });
-    
-    describe('onPlayerReady', ()=>{
+
+    describe('onPlayerReady', () => {
         it('Should set vgFor', () => {
             vgVol.vgFor = 'test';
             vgVol.onPlayerReady();
@@ -37,67 +37,67 @@ describe('Volume control', () => {
         });
     });
 
-    describe('onMouseDown', ()=>{
-        it('Should set isDragging to true', ()=>{
+    describe('onMouseDown', () => {
+        it('Should set isDragging to true', () => {
             vgVol.onMouseDown({clientX: 0});
             expect(vgVol.isDragging).toBe(true);
         });
-        it('Should set mouseDownPosX to event.x', ()=>{
+        it('Should set mouseDownPosX to event.x', () => {
             vgVol.onMouseDown({clientX: 99});
             expect(vgVol.mouseDownPosX).toBe(99);
         });
     });
-    
-    describe('onDrag', ()=>{
-        beforeEach(()=>{
+
+    describe('onDrag', () => {
+        beforeEach(() => {
             spyOn(vgVol, 'setVolume');
             spyOn(vgVol, 'calculateVolume');
         });
-        it('Should call setVolume when dragging', ()=>{
+        it('Should call setVolume when dragging', () => {
             vgVol.isDragging = true;
             vgVol.onDrag({clientX: 0});
             expect(vgVol.setVolume).toHaveBeenCalled();
         });
-        it('Should not call setVolume when not dragging', ()=>{
+        it('Should not call setVolume when not dragging', () => {
             vgVol.isDragging = false;
             vgVol.onDrag({clientX: 0});
             expect(vgVol.setVolume).not.toHaveBeenCalled();
         });
     });
 
-    describe('onStopDrag', ()=>{
-        beforeEach(()=>{
+    describe('onStopDrag', () => {
+        beforeEach(() => {
             spyOn(vgVol, 'setVolume');
             spyOn(vgVol, 'calculateVolume');
         });
-        it('Should toggle dragging value when dragging', ()=>{
+        it('Should toggle dragging value when dragging', () => {
             vgVol.isDragging = true;
             vgVol.mouseDownPosX = 0;
-            vgVol.onStopDrag({clientX:0});
+            vgVol.onStopDrag({clientX: 0});
             expect(vgVol.isDragging).toBe(false);
         });
-        it('Should call setVolume when dragging and x positions match', ()=>{
+        it('Should call setVolume when dragging and x positions match', () => {
             vgVol.isDragging = true;
             vgVol.mouseDownPosX = 0;
-            vgVol.onStopDrag({clientX:0});
+            vgVol.onStopDrag({clientX: 0});
             expect(vgVol.setVolume).toHaveBeenCalled();
         });
-        it('Should not call setVolume when dragging but x positions dont match', ()=>{
+        it('Should not call setVolume when dragging but x positions dont match', () => {
             vgVol.isDragging = true;
             vgVol.mouseDownPosX = 1;
-            vgVol.onStopDrag({clientX:0});
+            vgVol.onStopDrag({clientX: 0});
             expect(vgVol.setVolume).not.toHaveBeenCalled();
         });
-        it('Should not call setVolume when not dragging', ()=>{
+        it('Should not call setVolume when not dragging', () => {
             vgVol.isDragging = false;
             vgVol.mouseDownPosX = 0;
-            vgVol.onStopDrag({clientX:0});
+            vgVol.onStopDrag({clientX: 0});
             expect(vgVol.setVolume).not.toHaveBeenCalled();
         });
     });
 
-    describe('calculateVolume', ()=>{
-        it('Shoud calculate volume based on volumeBar position', ()=>{
+    describe('calculateVolume', () => {
+        it('Shoud calculate volume based on volumeBar position', () => {
             // mock volumeBarRef ViewChild
             vgVol.volumeBarRef = {
               nativeElement: {
@@ -110,8 +110,8 @@ describe('Volume control', () => {
         });
     });
 
-    describe('setVolume', ()=>{
-        it('Shoud convert volume to a value between 0 and 1', ()=>{
+    describe('setVolume', () => {
+        it('Shoud convert volume to a value between 0 and 1', () => {
             vgVol.target = {
                 volume: 3.33
             };
@@ -120,20 +120,20 @@ describe('Volume control', () => {
 
             vgVol.setVolume(110);
             expect(vgVol.target.volume).toBe(1);
-            
+
             vgVol.setVolume(-10);
             expect(vgVol.target.volume).toBe(0);
         });
     });
-    
-    describe('getVolume', ()=>{
-        it('Shoud get target volume when target is set', ()=>{
+
+    describe('getVolume', () => {
+        it('Shoud get target volume when target is set', () => {
             vgVol.target = {
                 volume: 3.33
             };
             expect(vgVol.getVolume()).toBe(3.33);
         });
-        it('Shoud get 0 volume when target is not set', ()=>{
+        it('Shoud get 0 volume when target is not set', () => {
             vgVol.target = undefined;
             expect(vgVol.getVolume()).toBe(0);
         });
