@@ -1,4 +1,4 @@
-import {Injectable} from '@angular/core';
+import { Injectable } from '@angular/core';
 
 @Injectable()
 export class VgUtils {
@@ -7,14 +7,14 @@ export class VgUtils {
      * https://gist.github.com/paulirish/211209
      * @returns {number}
      */
-    static getZIndex():number {
+    static getZIndex(): number {
         let zIndex = 1;
-        let elementZIndex:number;
+        let elementZIndex: number;
 
-        let tags = document.getElementsByTagName('*');
+        const tags = document.getElementsByTagName('*');
 
         for (let i = 0, l = tags.length; i < l; i++) {
-            elementZIndex = parseInt(window.getComputedStyle(tags[i])["z-index"], 10);
+            elementZIndex = parseInt(window.getComputedStyle(tags[i])['z-index'], 10);
 
             if (elementZIndex > zIndex) {
                 zIndex = elementZIndex + 1;
@@ -26,7 +26,7 @@ export class VgUtils {
 
     // Very simple mobile detection, not 100% reliable
     static isMobileDevice() {
-        return (typeof window.orientation !== "undefined") || (navigator.userAgent.indexOf("IEMobile") !== -1);
+        return (typeof window.orientation !== 'undefined') || (navigator.userAgent.indexOf('IEMobile') !== -1);
     }
 
     static isiOSDevice() {
