@@ -1,50 +1,52 @@
+// tslint:disable: variable-name
+// tslint:disable: no-namespace
 declare namespace google {
     export namespace ima {
-        
+
         /**
          * AdDisplayContainer
          */
         export class AdDisplayContainer {
-            constructor(containerElement:HTMLElement);
-            initialize():void;
+            constructor(containerElement: HTMLElement);
+            initialize(): void;
         }
 
         /**
          * AdsLoader
          */
         export class AdsLoader {
-            constructor(containerElement:AdDisplayContainer);
+            constructor(containerElement: AdDisplayContainer);
             contentComplete(): void;
-            addEventListener(type:string, callback:(evt:AdsManagerLoadedEvent)=>void, useCapture?:boolean): any;
-            requestAds(adsRequest:AdsRequest, opt_userRequestContext?:any):void;
+            addEventListener(type: string, callback: (evt: AdsManagerLoadedEvent) => void, useCapture?: boolean): any;
+            requestAds(adsRequest: AdsRequest, opt_userRequestContext?: any): void;
         }
-        
+
         /**
          * AdsManager
          */
         export class AdsManager {
-            resize(width:number, height:number, viewMode:ViewMode):void;
-            addEventListener(type:string, callback:(evt:AdsManagerLoadedEvent)=>void, useCapture?:boolean): any;
-            init(width:number, height:number, viewMode:ViewMode, opt_videoElement?:HTMLVideoElement):void;
-            start():void;
-            getAdSkippableState():boolean;
-            skip():void;
-            destroy():void;
-            getCuePoints():Array<number>;
+            resize(width: number, height: number, viewMode: ViewMode): void;
+            addEventListener(type: string, callback: (evt: AdsManagerLoadedEvent) => void, useCapture?: boolean): any;
+            init(width: number, height: number, viewMode: ViewMode, opt_videoElement?: HTMLVideoElement): void;
+            start(): void;
+            getAdSkippableState(): boolean;
+            skip(): void;
+            destroy(): void;
+            getCuePoints(): Array<number>;
         }
-        
+
         /**
          * AdsManagerLoadedEvent
          */
         class AdsManagerLoadedEventTypes {
-            ADS_MANAGER_LOADED;
+            ADS_MANAGER_LOADED: string;
         }
         export class AdsManagerLoadedEvent {
-            static Type:AdsManagerLoadedEventTypes;
+            static Type: AdsManagerLoadedEventTypes;
             getAdsManager(
-                contentPlayback:{currentTime:number, duration:number}, 
-                adsRenderingSettings?:AdsRenderingSettings
-            ):AdsManager;
+                contentPlayback: { currentTime: number, duration: number },
+                adsRenderingSettings?: AdsRenderingSettings
+            ): AdsManager;
         }
 
         /**
@@ -52,31 +54,31 @@ declare namespace google {
          */
         class AdsRenderingSettings {
         }
-        
+
         /**
          * AdEvent
          */
         class AdEventType {
-            CONTENT_PAUSE_REQUESTED;
-            CONTENT_RESUME_REQUESTED;
-            SKIPPABLE_STATE_CHANGED;
-            ALL_ADS_COMPLETED;
-            COMPLETE;
+            CONTENT_PAUSE_REQUESTED: string;
+            CONTENT_RESUME_REQUESTED: string;
+            SKIPPABLE_STATE_CHANGED: string;
+            ALL_ADS_COMPLETED: string;
+            COMPLETE: string;
         }
         export class AdEvent {
-            static Type:AdEventType;
+            static Type: AdEventType;
         }
-        
+
         /**
          * AdErrorEvent
          */
         class AdErrorEventTypes {
-            AD_ERROR;
+            AD_ERROR: string;
         }
         export class AdErrorEvent {
-            static Type:AdErrorEventTypes;
+            static Type: AdErrorEventTypes;
         }
-        
+
         /**
          * AdsRequest
          */
@@ -87,7 +89,7 @@ declare namespace google {
             nonLinearAdSlotWidth: number;
             nonLinearAdSlotHeight: number;
         }
-        
+
         /**
          * ViewMode
          */
@@ -100,28 +102,28 @@ declare namespace google {
 
 declare namespace googletag {
     export namespace cmd {
-        function push(command:Function):void;
+        function push(command: () => void): void;
     }
-    
+
     export class Service {
     }
-    
+
     export class CompanionAdsService extends Service {
-        setRefreshUnfilledSlots(value:boolean):void;
+        setRefreshUnfilledSlots(value: boolean): void;
     }
-    
+
     export class PubAdsService extends Service {
-        enableVideoAds():void;
+        enableVideoAds(): void;
     }
-    
+
     export class GeneralSize {
     }
-    
+
     export class Slot {
         addService(service: Service): Slot;
     }
 
-    function defineSlot(adUnitPath:string, size:GeneralSize, opt_div:string): Slot;
+    function defineSlot(adUnitPath: string, size: GeneralSize, opt_div: string): Slot;
     function companionAds(): CompanionAdsService;
     function pubads(): PubAdsService;
     function enableServices(): void;
