@@ -1,10 +1,10 @@
-import { NgModule }      from '@angular/core';
+import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
-import {VgImaAds} from './vg-ima-ads';
+import { VgImaAds } from './vg-ima-ads';
 
 @NgModule({
-    imports: [ CommonModule ],
+    imports: [CommonModule],
     declarations: [
         VgImaAds
     ],
@@ -12,4 +12,4 @@ import {VgImaAds} from './vg-ima-ads';
         VgImaAds
     ]
 })
-export class VgImaAdsModule {}
+export class VgImaAdsModule { }
