@@ -1,5 +1,5 @@
 describe('Videogular IMA Ads', () => {
     it('Should be ok', () => {
-        expect("ok").toBe("ok");
+        expect('ok').toBe('ok');
     });
 });
