@@ -1,10 +1,10 @@
-import { NgModule }      from '@angular/core';
+import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
-import {VgOverlayPlay} from './vg-overlay-play';
+import { VgOverlayPlay } from './vg-overlay-play';
 
 @NgModule({
-    imports: [ CommonModule ],
+    imports: [CommonModule],
     declarations: [
         VgOverlayPlay
     ],
@@ -12,4 +12,4 @@ import {VgOverlayPlay} from './vg-overlay-play';
         VgOverlayPlay
     ]
 })
-export class VgOverlayPlayModule {}
+export class VgOverlayPlayModule { }
